+++ src/main/util.ts
@@ -1,6 +1,7 @@
 /* eslint import/prefer-default-export: off, import/no-mutable-exports: off */
 import { URL } from 'url';
 import path from 'path';
+import { existsSync, mkdirSync, writeFileSync } from 'fs';
 
 export let resolveHtmlPath: (htmlFileName: string) => string;
 
@@ -16,3 +17,29 @@ if (process.env.NODE_ENV === 'development') {
     return `file://${path.resolve(__dirname, '../renderer/', htmlFileName)}`;
   };
 }
+
+export const padZeros = (num: number, len: number) => {
+  return String(num).padStart(len, '0');
+};
+
+export const integerDivide = (a: number, b: number) => {
+  return Math.floor(a / b);
+};
+
+export const writeFile = (filePath: string, data: any) => {
+  try {
+    writeFileSync(filePath, data, { flag: 'w' });
+  } catch (err) {
+    console.error(err);
+  }
+};
+
+export const mkdir = (dirPath: string) => {
+  if (!existsSync(dirPath)) {
+    try {
+      mkdirSync(dirPath);
+    } catch (err) {
+      console.error(err);
+    }
+  }
+};
